Restructured game module and added new endpoint
