+++ src/chess/game/dto/game.dto.ts
@@ -4,3 +4,8 @@ export class CreateGameLinkDto {
     @IsString()
     gameMode: string;
 }
+
+export class GetGameByGameLinkDto {
+    @IsString()
+    encodedId: string;
+}
